requirements.txt
Django==4.0.1
Jinja2==3.0.3
Pillow==9.0.0
django-allauth==0.48.0
django-mptt=0.13.4
django-admin-tools==0.9.2
django-debug-toolbar==3.2.4

flake8==4.0.1
black==21.12b0
braintree==4.14.0
